refactor(strand-report): merge ongoing and final assessment into single tab
Removes the Ongoing/Final sub-tab split in the Assessment view by
folding the ongoing assessment component into the assessment component.
The unified view shows moment-grouped assessment points followed by
the goals section on one page.

Removes the /ongoing_assessment routes and replaces the overlay URL
pattern with /assessment/assessment_point/:id. Drops the "goals
without assessment entries" collapsed section. Updates engagement
tracking to reflect three trackable tabs instead of four.
