+++ lib/lanttern_web/live/pages/shared/strand_report/strand_report_assessment_component.ex
@@ -16,20 +16,55 @@ defmodule LantternWeb.StrandReportLive.StrandReportAssessmentComponent do
   alias Lanttern.Assessments
   alias Lanttern.Assessments.AssessmentPoint
   alias Lanttern.GradesReports
+  alias Lanttern.LearningContext
   alias Lanttern.Reporting
   alias LantternWeb.Assessments.EntryParticleComponent
   alias LantternWeb.Assessments.StrandGoalDetailsOverlayComponent
+  alias LantternWeb.Assessments.StudentAssessmentPointDetailsOverlayComponent
   alias LantternWeb.Attachments.AttachmentViewComponent
   alias LantternWeb.GradesReports.GradeDetailsOverlayComponent
   alias LantternWeb.GradesReports.StudentGradesReportEntryButtonComponent
 
   import LantternWeb.AssessmentsComponents
+  import LantternWeb.ReportingComponents, only: [strand_report_assessment_point_card: 1]
 
   @impl true
   def render(assigns) do
     ~H"""
     <div class={@class}>
       <.responsive_container>
+        <h2 class="font-display font-black text-2xl">{gettext("Ongoing assessment")}</h2>
+        <p class="mt-4">
+          {gettext("Information about ongoing assessment points, grouped by moments.")}
+        </p>
+
+        <section id="ongoing-assessment-points" class="mt-10">
+          <div class="mt-6 space-y-10">
+            <div :for={moment <- @moments} id={"moment-#{moment.id}-ap-group"}>
+              <h4 class="font-display font-bold text-lg">{moment.name}</h4>
+              <div
+                id={"moment-#{moment.id}-sortable-aps"}
+                phx-update="stream"
+              >
+                <.strand_report_assessment_point_card
+                  :for={{dom_id, ap} <- @streams["moment_#{moment.id}_assessment_points"] || []}
+                  id={dom_id}
+                  assessment_point={ap}
+                  patch={"#{@base_path}/assessment_point/#{ap.id}"}
+                  class="mt-4"
+                />
+                <.empty_state_simple
+                  class="p-4 mt-4 hidden only:block"
+                  id={"moment-#{moment.id}-assessment-empty"}
+                >
+                  {gettext("No assessment points in this moment")}
+                </.empty_state_simple>
+              </div>
+            </div>
+          </div>
+        </section>
+      </.responsive_container>
+      <.responsive_container class="mt-16">
         <h2 class="font-display font-black text-2xl">{gettext("Goals assessment")}</h2>
         <p class="mt-4">
           {gettext("Final assessment points information, grouped by curriculum item.")}
@@ -116,35 +151,15 @@ defmodule LantternWeb.StrandReportLive.StrandReportAssessmentComponent do
             </.card_base>
           </div>
         </div>
-
-        <div :if={@has_strand_goals_without_student_entries} class="mt-10">
-          <div class="flex items-center gap-2">
-            <h4 class="flex-1 font-display font-black text-ltrn-subtle">
-              {gettext("Goals without assessment entries")}
-            </h4>
-            <.toggle_expand_button
-              id="toggle-strand-goals-without-student-entries"
-              target_selector="#strand-goals-without-student-entries"
-              initial_is_expanded={false}
-            />
-          </div>
-          <div id="strand-goals-without-student-entries" phx-update="stream" class="hidden">
-            <.goal_card
-              :for={
-                {dom_id, {goal, entry, moment_entries}} <-
-                  @streams.strand_goals_without_student_entries
-              }
-              id={dom_id}
-              patch={"#{@base_path}/strand_goal/#{goal.id}"}
-              goal={goal}
-              entry={entry}
-              moment_entries={moment_entries}
-              has_evidence={@goal_has_evidences_map[goal.id]}
-              prevent_preview={@prevent_final_assessment_preview}
-            />
-          </div>
-        </div>
       </.responsive_container>
+      <.live_component
+        :if={@assessment_point_id}
+        module={StudentAssessmentPointDetailsOverlayComponent}
+        id="ongoing-assessment-point-details-component"
+        assessment_point_id={@assessment_point_id}
+        student_id={@student_report_card.student_id}
+        on_cancel={JS.patch(@base_path)}
+      />
       <.live_component
         :if={@strand_goal_id}
         module={StrandGoalDetailsOverlayComponent}
@@ -322,15 +337,14 @@ defmodule LantternWeb.StrandReportLive.StrandReportAssessmentComponent do
     socket =
       socket
       |> assign(:class, nil)
+      |> assign(:moments, [])
+      |> assign(:assessment_points_ids, [])
+      |> assign(:assessment_point_id, nil)
       |> assign(:strand_goal_id, nil)
       |> assign(:student_grades_report_entry_id, nil)
       |> assign(:initialized, false)
       |> stream_configure(
         :strand_goals_student_entries,
-        dom_id: fn {goal, _, _} -> "goal-#{goal.id}" end
-      )
-      |> stream_configure(
-        :strand_goals_without_student_entries,
         dom_id: fn {goal, _, _} -> "goal-#{goal.id}" end
       )
       |> stream_configure(
@@ -347,6 +361,7 @@ defmodule LantternWeb.StrandReportLive.StrandReportAssessmentComponent do
       socket
       |> assign(assigns)
       |> initialize()
+      |> assign_assessment_point_id()
       |> assign_strand_goal_id()
       |> assign_student_grades_report_entry_id()
 
@@ -355,6 +370,7 @@ defmodule LantternWeb.StrandReportLive.StrandReportAssessmentComponent do
 
   defp initialize(%{assigns: %{initialized: false}} = socket) do
     socket
+    |> load_moments_and_assessment_points()
     |> stream_strand_goals_student_entries()
     |> assign_prevent_final_assessment_preview()
     |> stream_strand_evidences()
@@ -381,23 +397,12 @@ defmodule LantternWeb.StrandReportLive.StrandReportAssessmentComponent do
         not is_nil(entry) or moments_entries != []
       end)
 
-    strand_goals_without_student_entries =
-      all_strand_goals_student_entries
-      |> Enum.filter(fn {_, entry, moments_entries} ->
-        is_nil(entry) and moments_entries == []
-      end)
-
     socket
     |> stream(:strand_goals_student_entries, strand_goals_student_entries)
     |> assign(:strand_goals_ids, strand_goals_ids)
-    |> stream(:strand_goals_without_student_entries, strand_goals_without_student_entries)
     |> assign(
       :has_strand_goals_with_student_entries,
       strand_goals_student_entries != []
-    )
-    |> assign(
-      :has_strand_goals_without_student_entries,
-      strand_goals_without_student_entries != []
     )
   end
 
@@ -500,4 +505,48 @@ defmodule LantternWeb.StrandReportLive.StrandReportAssessmentComponent do
 
   defp assign_student_grades_report_entry_id(socket),
     do: assign(socket, :student_grades_report_entry_id, nil)
+
+  defp load_moments_and_assessment_points(socket) do
+    strand_id = socket.assigns.strand_report.strand_id
+
+    moments = LearningContext.list_moments(strands_ids: [strand_id])
+    moments_ids = Enum.map(moments, & &1.id)
+
+    assessment_points =
+      Assessments.list_strand_moments_assessment_points_with_student_entries(
+        socket.assigns.current_scope,
+        socket.assigns.student_report_card.student,
+        socket.assigns.strand_report.strand_id
+      )
+
+    socket
+    |> assign(:moments, moments)
+    |> assign(:moments_ids, moments_ids)
+    |> assign(:assessment_points_ids, Enum.map(assessment_points, &"#{&1.id}"))
+    |> stream_assessment_points_by_moment(assessment_points, moments)
+  end
+
+  defp stream_assessment_points_by_moment(socket, assessment_points, moments) do
+    Enum.reduce(moments, socket, fn moment, socket ->
+      moment_aps =
+        assessment_points
+        |> Enum.filter(&(&1.moment_id == moment.id))
+
+      stream(socket, "moment_#{moment.id}_assessment_points", moment_aps)
+    end)
+  end
+
+  defp assign_assessment_point_id(
+         %{assigns: %{params: %{"assessment_point_id" => assessment_point_id}}} = socket
+       ) do
+    # simple guard to prevent viewing details from unrelated assessment points
+    assessment_point_id =
+      if assessment_point_id in socket.assigns.assessment_points_ids do
+        assessment_point_id
+      end
+
+    assign(socket, :assessment_point_id, assessment_point_id)
+  end
+
+  defp assign_assessment_point_id(socket), do: assign(socket, :assessment_point_id, nil)
 end
